+++ src/js/components/Header/Header.less
@@ -1,46 +1,47 @@
 @import (reference) '../../../styles/config.less';
 
-nav {
-  &.navbar-default {
-    background-color: transparent;
-    border-color: transparent;
-    .transition(background 0.2s linear);
-    .transition(opacity 0.4s linear);
-    border-radius: 0;
-    z-index: @nav-z + 1;
-    .navbar-nav > li > a {
-      color: @brand-primary;
-      background-color: fade(@brand-primary, 10%);
-      &:focus {
-        background-color: fade(@brand-primary, 10%);
-        color: @brand-primary;
-      }
-      &:hover {
-        color: @color-indigo;
-        background-color: @color-yellow;
-      }
-    }
-    .navbar-right {
-      margin-right: 15px;
-      .isMobile({ margin-right: @app-padding-mobile; });
-    }
-  }
-
-  .navbar-collapse {
-    border: none;
-    top: 5px;
-  }
+.navbar.navbar-default {
+  background-color: transparent;
+  border-color: transparent;
+  .transition(background 0.2s linear);
+  .transition(opacity 0.4s linear);
+  border-radius: 0;
+  z-index: @nav-z + 1;
   &.navbar-fixed-color {
     background-color: @color-indigo;
     background-color: fade(@color-indigo, 90%);
   }
-  .navbar-brand {
-    padding: 0;
-    .isMobile({ width: 9em; padding-top: 10px; });
-    img {
-      .isMobile({ width: 100%; });
+  .navbar-nav > li > a {
+    color: @brand-primary;
+    background-color: fade(@brand-primary, 10%);
+    &:focus {
+      background-color: fade(@brand-primary, 10%);
+      color: @brand-primary;
+    }
+    &:hover {
+      color: @color-indigo;
+      background-color: @color-yellow;
     }
   }
+  .navbar-right {
+    margin: 2px 0 0 0;
+  }
+
+  nav {
+    .navbar-collapse {
+      border: none;
+      top: 5px;
+    }
+
+    .navbar-brand {
+      padding: 0;
+      .isMobile({ width: 9em; });
+      img {
+        .isMobile({ width: 100%; padding: 5px; });
+      }
+    }
+  }
+
   .collapse.in {
     background-color: @color-indigo;
   }
@@ -52,27 +53,27 @@ nav {
     color: @brand-primary;
     text-transform: capitalize;
   }
-}
 
-.login-button {
-  float: right !important;
-  color: #000000;
-  background-color: #E4BD00;
-  border: none;
-  border-radius: 20px;
-  font-size: 16px;
-  line-height: 24px;
-  padding-left: 16px;
-  padding-right: 16px;
-  margin: 20px 25px 0 0;
+  .login-button {
+    float: right !important;
+    color: #000000;
+    background-color: #E4BD00;
+    border: none;
+    border-radius: 20px;
+    font-size: 16px;
+    line-height: 24px;
+    padding-left: 16px;
+    padding-right: 16px;
+    margin: 20px 25px 0 0;
 
-  @media all and (max-width: @screen-sm-min) {
-    margin: 10px 10px 0 0;
+    @media all and (max-width: @screen-sm-min) {
+      margin: 10px 10px 0 0;
+    }
   }
 }
 
 @media (min-width: 768px) {
-  nav {
+  .navbar {
     .navbar-header {
       width: 100%;
       text-align: center;
