+++ src/js/components/Movies/MoviesSlider.less
@@ -33,8 +33,8 @@
     width: 100%;
     overflow: hidden;
     &.axis-x {
-      display: inline-block;
-      white-space: nowrap;
+      //display: inline-block;
+      //white-space: nowrap;
       > div {
         > div > div > div {
           width: 100%;
