+++ src/js/components/Movies/UserMoviesList.less
@@ -13,7 +13,7 @@
         white-space: nowrap;
         > div {
           > div > div > div {
-            height: @thumb-height;
+            height: @thumb-height - (50px);
           }
         }
       }
