+++ src/js/components/Search/SearchBox.less
@@ -2,7 +2,7 @@
 
 .search-box {
   @box-size: 200px;
-  margin: 2px 2px 2px 10px;
+  margin: 2px 2px 2px 2px;
   color: @brand-primary;
   border-radius: @btn-radius;
   text-align: center;
@@ -12,7 +12,7 @@
     font-family: 'FontAwesome';
     font-size: 1.8em;
     line-height: 2.2;
-    left: 22px;
+    left: 14px;
     color: @brand-primary;
   }
 
@@ -33,7 +33,7 @@
       display: block;
     }
     &:before {
-      .isMobile({ left: 22px; });
+      .isMobile({ left: 14px; });
     }
     input {
       .isMobile({ width: 100%; });
